CTM: clean up linear solver
