+++ src/CTM/CMakeLists.txt
@@ -4,7 +4,7 @@ CTM_TemperatureEvaluator.cpp
 CTM_ThermalProblem.cpp
 CTM_MechanicsProblem.cpp
 CTM_Application.cpp
-linear_solver.cpp
+CTM_LinearSolver.cpp
 CTM_Solver.cpp)
 
 include_directories(../LCM/problems)
